Minor breaking change: process quickfix output in chunks
Previosly we used default Plenary.Job callback which was called on every
lines. But it was very slow. So now we process data in chunks which
results in significant speedup. But this means that `on_build_output` now
accepts list of string, so this is a minor breaking change for people
who use this setting.